Fix Conflict with new database

+++ swp391_debo_be/Services/Implements/TreatmentService.cs
@@ -4,17 +4,29 @@ using swp391_debo_be.Cores;
 using swp391_debo_be.Dto.Implement;
 using swp391_debo_be.Repository.Implement;
 using swp391_debo_be.Services.Interfaces;
+using System.Collections.Generic;
 using System.Net;
 
 namespace swp391_debo_be.Services.Implements
 {
-    public class TreatmentService
+    public class TreatmentService : ITreatmentService
     {
         public async Task<ApiRespone> addTreatmentAsync(TreatmentDto treatment)
         {
             var response = new ApiRespone();
             try
             {
+                // -1 mean get all treatment 
+                var existingTreatments = await CTreatment.getAllTreatmentAsync(1, -1);
+
+                // Check if the treatment ID already exists
+                if (existingTreatments.Any(t => t.Id == treatment.Id))
+                {
+                    response.StatusCode = HttpStatusCode.BadRequest;
+                    response.Success = false;
+                    response.Message = "ID cannot be duplicated";
+                    return response;
+                }
                 var newTreat = await CTreatment.addTreatmentAsync(treatment);
                 var treat = await CTreatment.getTreatmentAsync(newTreat);
                 response.StatusCode = HttpStatusCode.OK;
@@ -36,12 +48,22 @@ namespace swp391_debo_be.Services.Implements
             var response = new ApiRespone();
             try
             {
-                var treat = await CTreatment.getTreatmentAsync(id);
-                await CTreatment.deleteTreatmentAsync(id);
-                response.StatusCode = HttpStatusCode.OK;
-                response.Data = treat;
-                response.Success = true;
-                response.Message = "Treatment data is deleted successfully.";
+                var existingTreatments = await CTreatment.getTreatmentAsync(id);
+                if (existingTreatments != null)
+                {
+                    await CTreatment.deleteTreatmentAsync(id);
+                    response.StatusCode = HttpStatusCode.OK;
+                    response.Data = existingTreatments;
+                    response.Success = true;
+                    response.Message = "Treatment data is deleted successfully.";
+                }
+                else
+                {
+                    response.StatusCode = HttpStatusCode.NotFound;
+                    response.Success = false;
+                    response.Message = "Treatment not found.";
+                }
+
             }
             catch (Exception ex)
             {
@@ -57,7 +79,7 @@ namespace swp391_debo_be.Services.Implements
             var response = new ApiRespone();
             try
             {
-                var data = await CTreatment.getAllTreatmentAsync(page,limit);
+                var data = await CTreatment.getAllTreatmentAsync(page, limit);
                 response.StatusCode = HttpStatusCode.OK;
                 response.Data = data;
                 response.Success = true;
@@ -77,11 +99,20 @@ namespace swp391_debo_be.Services.Implements
             var response = new ApiRespone();
             try
             {
-                var data = await CTreatment.getTreatmentAsync(id);
-                response.StatusCode = HttpStatusCode.OK;
-                response.Data = data;
-                response.Success = true;
-                response.Message = "Treatment data retrieved successfully.";
+                var existingTreatments = await CTreatment.getTreatmentAsync(id);
+                if (existingTreatments != null)
+                {
+                    response.StatusCode = HttpStatusCode.OK;
+                    response.Data = existingTreatments;
+                    response.Success = true;
+                    response.Message = "Treatment data retrieved successfully.";
+                }
+                else
+                {
+                    response.StatusCode = HttpStatusCode.NotFound;
+                    response.Success = false;
+                    response.Message = "Treatment not found.";
+                }
             }
             catch (Exception ex)
             {
@@ -103,7 +134,7 @@ namespace swp391_debo_be.Services.Implements
                     response.Success = false;
                     response.Message = "This Treatment does not exist in system";
                 }
-                await CTreatment.updateTreatmentAsync(id,treatment);
+                await CTreatment.updateTreatmentAsync(id, treatment);
                 var data = await CTreatment.getTreatmentAsync(id);
                 response.StatusCode = HttpStatusCode.OK;
                 response.Data = data;
